Track python structure in rust
